Fix [start, end] with negative numbers (#52)
The range was not working as intended when the end was less than the start. As a developer I expect the first number to be inclusive and the second number to be exclusive, regardless of negative sign. This fixes the code to match that behavior. However, for panels the fix is such that XY, XZ, and YZ creators use positive axes, this doesn't fix panels that use negative axes.

Close https://github.com/villagekit/villagekit/issues/47